% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCohort.R
\name{addCohort}
\alias{addCohort}
\title{Add another cohort to the profile}
\usage{
addCohort(massPools = data.frame(age = 0, fast_OM = 0, slow_OM = 0,
  mineral = 0, root_mass = 0, layer_top = 0, layer_bottom = 0), ssc,
  meanTidalHeight, rootTurnover, rootOmFrac, omDecayRate,
  totalRootMass_per_area, rootDepthMax, consts, dt_yr = 1)
}
\arguments{
\item{massPools}{}

\item{ssc}{}

\item{meanTidalHeight}{}

\item{rootTurnover}{}

\item{rootOmFrac}{}

\item{omDecayRate}{}

\item{totalRootMass_per_area}{}

\item{rootDepthMax}{}

\item{consts}{}

\item{dt_yr}{}
}
\description{
Add another cohort to the profile
}
