% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateCohorts.R
\name{animateCohorts}
\alias{animateCohorts}
\title{Animate sediment cohorts}
\usage{
animateCohorts(cohorts, scenario, filename = "MEM-CTM-animated.gif",
  savePath = getwd(), chPalette = c("#56B4E9", "#999999", "#E69F00",
  "#009E73"), trackThresholds = c("MSL", "MHW"), duration = 30,
  width = 4.5, height = 4.5)
}
\arguments{
\item{cohorts}{data frame, annually tracked soil mass cohorts output from runMemWithCohorts}

\item{scenario}{data frame, annual summaries of inputs and outputs from runMemWithCohorts}

\item{filename, }{character, name of the output file}

\item{savePath}{character, filepath to save animation to}

\item{trackThresholds}{vector, a vector of characters indicating which water leves in the scenario table to map as horizontal lines}

\item{duration}{numeric, length in seconds of the animation}

\item{width}{numeric, width in inches of the .gif}

\item{height}{numeric, height in inches of the .gif}

\item{chPallette}{vector, a vector of colors to use to symbolize the different mass cohorts}
}
\description{
This function takes two MEM outputs, cohorts and scenario tables, as inputs and visualizes soil formation as an animated .gif
}
