% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateCohortsAccrossTransect.R
\name{animateCohortsAccrossTransect}
\alias{animateCohortsAccrossTransect}
\title{Animate sediment cohorts accross an elevation transect}
\usage{
animateCohortsAccrossTransect(scenarioTransect, cohortsTransect,
  duration = 30, width = 5, height = 3, savePath = getwd(),
  filename = "MEM-transect-animation.gif")
}
\arguments{
\item{scenarioTransect}{data frame, annual summaries of inputs and outputs from runMemWithCohortsAccrossTransect}

\item{cohortsTransect}{data frame, annually tracked soil mass cohorts output from runMemWithCohortsAccrossTransect}

\item{duration}{numeric, length in seconds of the animation}

\item{width}{numeric, width in inches of the .gif}

\item{height}{numeric, height in inches of the .gif}

\item{savePath}{character, filepath to save animation to}

\item{filename, }{character, name of the output file}
}
\description{
This function takes two runMemWithCohortsAccrossTransect outputs, cohortsTransect and scenarioTransect tables, as inputs and visualizes soil formation as an animated .gif
}
