% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHighTideScenario.R
\name{buildHighTideScenario}
\alias{buildHighTideScenario}
\title{Build high tides into scenario table}
\usage{
buildHighTideScenario(
  scenarioCurve,
  meanSeaLevelDatum = scenarioCurve$meanSeaLevel[1],
  meanHighWaterDatum,
  meanHighHighWaterDatum,
  meanHighHighWaterSpringDatum,
  lunarNodalAmp
)
}
\arguments{
\item{scenarioCurve}{a data frame, including year, mean sea level, and suspended sediment concentration, with each row being a year in the scenario}

\item{meanSeaLevelDatum}{a numeric, Mean Sea Level over a tidal datum period}

\item{meanHighWaterDatum}{a numeric, Mean High Water level over the last datum period}

\item{meanHighHighWaterDatum}{a numeric, Mean Higher High Water level over the last datum period}

\item{meanHighHighWaterSpringDatum}{a numeric, Mean Higher High Spring Tide Water level over the last datum period}

\item{lunarNodalAmp}{the amplitude of the 18-year lunar nodal cycle}
}
\value{
a data frame, including the sea-level and suspended sediment concentraiton scenario inputted, with annual high tide datum(s) added
}
\description{
This function takes a sea-level scenario table and elevations for 1 to 3 tidal cycles and outputs a scenario table with annualized high tidel levels.
}
