% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScenarioCurve.R
\name{buildScenarioCurve}
\alias{buildScenarioCurve}
\title{Build scenario sea-level curve}
\usage{
buildScenarioCurve(startYear, endYear = startYear + 99, MSL,
  rslr0 = 0.3, rslrTotal = 100, ssc)
}
\arguments{
\item{startYear}{an integer, year in form YYYY, the start year of the scenario}

\item{endYear}{an integer, year in form YYYY, the end year of the scenario}

\item{MSL}{a numeric or a vector of numbers, either indicating mean sea-level at the start of scenario, or mean sea-level at each year of the sceanrio}

\item{rslr0}{a numeric, initial rate of relative sea-level rise}

\item{rslrTotal}{a numeric, total relative sea-level rise over the course of the scanario}

\item{ssc}{a numeric or a vector of numbers, either average annual suspended sediment concentration, or a vector of annual suspended sediment concentration for each year of the scenario}
}
\value{
a data frame including columns for year, sea-level, and suspended sediment concentration, and rows for each year in the scenario
}
\description{
This function builds an annualied set of MEM inputs including sea-level rise and suspended sediment concentrations.
}
\references{
Sweet, W. V., Kopp, R. E., Weaver, C. P., Obeysekera, J., Horton, R. M., Thieler, E. R., & Zervas, C. (2017). Global and regional sea level rise scenarios for the United States.

Intergovernmental Panel on Climate Change (2013). Summary for policymakers, in Climate Change 2013: The Physical Science Basis, edited by T. F. Stocker, D. Qin, G.-K. Plattner, M. Tignor, S. K. Allen, J. Boschung, A. Nauels, Y. Xia, V. Bex, and P. Midgley, pp. 3–29, Cambridge Univ. Press, Cambridge, U. K.
}
