% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCarbonFlux.R
\name{calculateCarbonFlux}
\alias{calculateCarbonFlux}
\title{Calculate Carbon Flux from cohorts and scenario tables}
\usage{
calculateCarbonFlux(cohorts, scenario, omToOcParams = list(B0 = 0, B1 = 0.48))
}
\arguments{
\item{cohorts}{a data frame, output from runCohortMem, tracking mineral and organic mass cohorts over each year of the simulation}

\item{scenario}{a data frame, output from runCohortMem, tracking annualized variables}

\item{omToOcParams}{a list, of numerics listed B0, B1 and optionally B3. These define a linear or parabolic relationship between organic matter and organic}
}
\value{
a list of data frames, including the annualized summaries, and mapped cohorts tracked for every year of the simulation.
}
\description{
Calculate Carbon Flux from cohorts and scenario tables
}
