% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertProfile_AgeToDepth.R
\name{convertProfile_AgeToDepth}
\alias{convertProfile_AgeToDepth}
\title{Convert soil profile from age to depth based cohorts}
\usage{
convertProfile_AgeToDepth(ageCohort, layerTop, layerBottom)
}
\arguments{
\item{ageCohort}{a dataframe specifying the age cohort. Must contain \code{layer_top}, \code{layer_bottom}, \code{age}, \code{fast_OM}, \code{slow_OM}, \code{mineral}, and \code{root_mass}.}

\item{layerTop}{a vector of the tops of the soil layers we want to convert to.}

\item{layerBottom}{a vector of the bottom of the soil layers we want to convert to.}
}
\value{
a data frame specifying the depth cohoprts. Contains \code{layer_top}, \code{layer_bottom}, \code{age}, \code{fast_OM}, \code{slow_OM}, \code{mineral}, and \code{root_mass}.
}
\description{
This function takes an age cohort soil profile and converts it to a depth based cohort using weighted sums and means.
}
