% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertZStarToZ.R
\name{convertZStarToZ}
\alias{convertZStarToZ}
\title{Convert from Z* to elevation}
\usage{
convertZStarToZ(zStar, meanHighWater, meanSeaLevel)
}
\arguments{
\item{zStar}{a numeric describing the normalized Z* elevation}

\item{meanHighWater}{a numeric, mean high water tide}

\item{meanSeaLevel}{a numeric, mean sea level}
}
\value{
a numerical describing the marsh elevation
}
\description{
Convert from Z* to elevation
}
