% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveredSedimentSimple.R
\name{deliveredSedimentSimple}
\alias{deliveredSedimentSimple}
\title{Calculate delivered sediment, simple method}
\usage{
deliveredSedimentSimple(z, ssc, MSL, MHW, MLW = MSL - MHW,
  settlingVelocity, k = 1, nTides = 704)
}
\arguments{
\item{z}{a numeric, elevation of the marsh}

\item{ssc}{a numeric, suspended sediment concentration of the water column}

\item{MSL}{a numeric, Mean Sea Level}

\item{MHW}{a numeric, Mean High Water level}

\item{MLW}{a numeric, Mean Low Water level}

\item{settlingVelocity}{a numeric, the number of times a water column will clear per tidal cycle}

\item{k}{a numeric, the fraction of available sediment captured by the marsh}

\item{nTides}{an integer, the number of flooding events in a year}

\item{floodPct}{a numeric, fraction of time per tidal cycle the marsh is inundated}
}
\value{
a numeric, the sediment delivered over the course of a year
}
\description{
This function calculates the delivered sediment for a marsh elevation over a year using the simplest possible approach.
}
