% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverSediment.R
\name{deliverSediment}
\alias{deliverSediment}
\title{Calculate delivered sediment over one or more types of tides}
\usage{
deliverSediment(
  z,
  suspendedSediment,
  nFloods = 705.79,
  meanSeaLevel,
  meanHighWater,
  meanHighHighWater = NA,
  meanHighHighWaterSpring = NA,
  meanLowWater = meanSeaLevel - meanHighWater,
  meanLowLowWater = meanSeaLevel - meanHighHighWater,
  meanLowLowWaterSpring = meanSeaLevel - meanHighHighWaterSpring,
  captureRate,
  floodTime.fn = floodTimeLinear
)
}
\arguments{
\item{z}{a numeric, elevation of the marsh}

\item{suspendedSediment}{a numeric, suspended sediment concentration of the water column}

\item{nFloods}{a numeric, the number of tidal flooding events per year}

\item{meanSeaLevel}{a numeric, Mean Sea Level}

\item{meanHighWater}{a numeric, Mean High Water level}

\item{meanHighHighWater}{a numeric, Mean Higher High Water level}

\item{meanHighHighWaterSpring}{a numeric, Mean Higher High Spring Tide Water level}

\item{meanLowWater}{a numeric, Mean Low Water level}

\item{meanLowLowWater}{a numeric, Mean Lower Low Water level}

\item{meanLowLowWaterSpring}{a numeric, Mean Low Lower Spring Tide Water level}

\item{captureRate}{a numeric, the number of times a water column will clear per tidal cycle}

\item{floodTime.fn}{a function, specify the method used to calculate flooding time per tidal cycle}
}
\value{
a numeric, the sediment delivered over the course of a year
}
\description{
This function calculates the delivered sediment for a marsh elevation by summarising between one and three types of tidal cycles.
}
