% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverSediment3TidalCycle.R
\name{deliverSediment3TidalCycle}
\alias{deliverSediment3TidalCycle}
\title{Calculate delivered sediment, three tidal cycle method}
\usage{
deliverSediment3TidalCycle(z, suspendedSediment, meanSeaLevel,
  meanHighWater, meanHighHighWater, meanHighHighWaterSpring = NA,
  meanLowWater = meanSeaLevel - meanHighWater,
  meanLowLowWater = meanSeaLevel - meanHighHighWater,
  meanLowLowWaterSpring = meanSeaLevel - meanHighHighWaterSpring,
  settlingVelocity, capturedSediment = 1)
}
\arguments{
\item{z}{a numeric, elevation of the marsh}

\item{suspendedSediment}{a numeric, suspended sediment concentration of the water column}

\item{meanSeaLevel}{a numeric, Mean Sea Level}

\item{meanHighWater}{a numeric, Mean High Water level}

\item{meanHighHighWater}{a numeric, Mean Higher High Water level}

\item{meanHighHighWaterSpring}{a numeric, Mean Higher High Spring Tide Water level}

\item{meanLowWater}{a numeric, Mean Low Water level}

\item{meanLowLowWater}{a numeric, Mean Lower Low Water level}

\item{meanLowLowWaterSpring}{a numeric, Mean Low Lower Spring Tide Water level}

\item{settlingVelocity}{a numeric, the number of times a water column will clear per tidal cycle}

\item{capturedSediment}{a numeric, the fraction of available sediment captured by the marsh}
}
\value{
a numeric, the sediment delivered over the course of a year
}
\description{
This function calculates the delivered sediment for a marsh elevation over by summarising over three types of tidal cycles.
}
