% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverSedimentFlexibly.R
\name{deliverSedimentFlexibly}
\alias{deliverSedimentFlexibly}
\title{Deliver sediment flexibly}
\usage{
deliverSedimentFlexibly(
  z,
  suspendedSediment,
  meanSeaLevel,
  meanHighWater,
  meanHighHighWater = NA,
  meanHighHighWaterSpring = NA,
  settlingVelocity,
  ...
)
}
\arguments{
\item{z}{a numeric, elevation of the marsh}

\item{suspendedSediment}{a numeric, suspended sediment concentration of the water column}

\item{meanSeaLevel}{a numeric, Mean Sea Level}

\item{meanHighWater}{a numeric, Mean High Water level}

\item{meanHighHighWater}{a numeric, Mean Higher High Water level}

\item{meanHighHighWaterSpring}{a numeric, Mean Higher High Spring Tide Water level}

\item{settlingVelocity}{a numeric, the number of times a water column will clear per tidal cycle}

\item{...}{}
}
\value{
a numeric, sediment delivered in a year of flooding
}
\description{
Function decides which sediment module to use based on available inputs
}
