% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthOfNonRootVolume.R
\name{depthOfNonRootVolume}
\alias{depthOfNonRootVolume}
\title{Depth of the given soil volume}
\usage{
depthOfNonRootVolume(nonRootVolume.arr, massLiveRoots.fn = NULL,
  totalRootMass_per_area, rootDepthMax, rootDensity, shape = "linear",
  soilLength = 1, soilWidth = 1, relTol = 1e-04, verbose = FALSE,
  ...)
}
\arguments{
\item{nonRootVolume.arr}{a numerical array of the cumulative non-rooting volume, assumed to be monotonically increasing}

\item{massLiveRoots.fn}{a function defining the mass of the live roots, if shape is not linear then this must be defined.}

\item{totalRootMass_per_area}{an integer that is the total mass per area of the roots, generally in g cm-3}

\item{rootDepthMax}{an integer that is the maximum root depth, generally in cm}

\item{rootDensity}{a numeric that is the root density in g per cm3}

\item{shape}{a character of the shape of the root distribution, currently only 'linear' is implimented}

\item{soilLength}{a numeric of the unit length of soil volume, generally 1 cm}

\item{soilWidth}{a numeric of the unitl witdth of the soil colume, generally 1cm}

\item{relTol}{a numeric the relative tolerence accepted for the general depth solution}

\item{verbose}{a boolean flag, true prints out a lot of strings to help debug the general solution}

\item{...}{}
}
\value{
A numeric corresponding to the depth of the specificed non root volume
}
\description{
Given a shape and mass for the roots, calculate the depth of a specific non-root volume. A specific solution is implimented for linear root volumes, and a general solution otherwise requires the \code{massLiveRoots.fn} to be specified. The general solution is a bineary search algorithm.
}
