% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthOfNonRootVolume.R
\name{depthOfNotRootVolume}
\alias{depthOfNotRootVolume}
\title{Depth of the given soil volume}
\usage{
depthOfNotRootVolume(nonRootVolume, totalRootMass_per_area, rootDepthMax,
  rootDensity, soilLength = 1, soilWidth = 1, shape = "linear", ...)
}
\arguments{
\item{nonRootVolume}{a numerical value (possibly array) that is the volume of interest, generally in cm3.}

\item{totalRootMass_per_area}{an integer that is the total mass per area of the roots, generally in g cm-3}

\item{rootDepthMax}{an integer that is the maximum root depth, generally in cm}

\item{rootDensity}{a numeric that is the root density in g per cm3}

\item{soilLength}{a numeric of the unit length of soil volume, generally 1 cm}

\item{soilWidth}{a numeric of the unitl witdth of the soil colume, generally 1cm}

\item{shape}{a character of the shape of the root distribution, currently only 'linear' is implimented}

\item{...}{}
}
\value{
A numeric (possibly array) corresponding to the depth of the specificed non root volume
}
\description{
Given a shape and mass for the roots, calculate the depth of a specific non-root volume.
}
