% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineInitialCohorts.R
\name{determineInitialCohorts}
\alias{determineInitialCohorts}
\title{Determine the initial conditions for a set of cohorts}
\usage{
determineInitialCohorts(initElv, MSL, MHW, MHHW = NA, MHHWS = NA, ssc,
  bMax, zVegMin, zVegMax, zVegPeak, plantElevationType, rootToShoot,
  rootTurnover, rootDepthMax, shape = "linear", omDecayRate,
  recalcitrantFrac, settlingVelocity, omPackingDensity = 0.085,
  mineralPackingDensity = 1.99, rootPackingDensity = omPackingDensity,
  initialCohorts = NA, uplandCohorts = NA, supertidalCohorts = NA,
  supertidalSedimentInput = NA)
}
\arguments{
\item{initElv}{a numeric, the initial elevation of the marsh at the start of the scenario}

\item{MSL}{a numeric, Mean Sea Level over the last datum period}

\item{MHW}{a numeric, Mean High Water level over the last datum period}

\item{MHHW}{a numeric (optional), Mean Higher High Water level over the last datum period}

\item{MHHWS}{a numeric (optional), Mean Higher High Spring Tide Water level over the last datum period}

\item{ssc}{a numeric, suspended sediment concentration of the water column}

\item{bMax}{a numeric, maximum biomass}

\item{zVegMin}{a numeric, lower elevation of biomass limit}

\item{zVegMax}{a numeric, upper elevation of biomass limit}

\item{zVegPeak}{a numeric, (optional) elevation of peak biomass}

\item{plantElevationType}{character, either "orthometric" or "dimensionless", specifying elevation reference of the vegetation growing elevations}

\item{rootToShoot}{a numeric, root to shoot ratio}

\item{rootTurnover}{a numeric, below ground biomass annual turnover rate}

\item{rootDepthMax}{a numeric, maximum (95\%) rooting depth}

\item{shape}{a character, "linear" or "exponential" describing the shape of the relationship between depth and root mass}

\item{omDecayRate}{a numeric, annual fractional mass lost}

\item{recalcitrantFrac}{a numeric, fraction of organic matter resistant to decay}

\item{settlingVelocity}{a numeric, the number of times a water column will clear per tidal cycle}

\item{omPackingDensity}{a numeric, the bulk density of pure organic matter}

\item{mineralPackingDensity}{a numeric, the bulk density of pure mineral matter}

\item{rootPackingDensity}{a numeric, the bulk density of pure root matter}

\item{initialCohorts}{a data frame, (optional) custom set of mass cohorts that will override any decision making this function does}

\item{uplandCohorts}{a data frame,  (optional) custom set of mass cohorts to be used if intiial elevation is higher than both maximum tidal height and maximum wetland vegetation tolerance}

\item{supertidalSedimentInput, }{a numeric, (optional) grams per cm^2 per year, an optional parameter which will define annual suspended sediment delivery to a sediment column that is is higher than maximum tidal height, but not maximum wetland vegetation tolerance}

\item{superTidalCohorts}{a data frame, (optional) custom set of mass cohorts to be used if intiial elevation is higher than maximum tidal height, but not maximum wetland vegetation tolerance}
}
\value{
a data frame of mass pools representing the initial conditions for a simulation
}
\description{
Determine the initial conditions for a set of cohorts
}
