% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillDepthCellsWithRoots.R
\name{fillDepthCellsWithRoots}
\alias{fillDepthCellsWithRoots}
\title{Calculates a root mass depth profile}
\usage{
fillDepthCellsWithRoots(depthMins = 0:149, depthMaxs = 1:150,
  totalRootBmass = 3000, rootShape = "linear", rootDepthMax = 30,
  customFunction = function(x) {     exp(-((x -
  maxRootDepth)^2/lambdaRoot^2)) }, customParams = list(lambdaRoot = 33,
  maxRootDepth = 30))
}
\arguments{
\item{depthMins}{}

\item{depthMaxs}{}

\item{totalRootBmass}{}

\item{rootShape}{}

\item{rootDepthMax}{}

\item{customFunction}{}

\item{customParams}{}
}
\description{
It will take any root biomass and maximum rooting depth, 
as well as vectors of minimum and maximum depths.
We use predefined equations if rootShape is set to linear or exponential
It can also take a custom function of x, as long as you define parameters in
a custom parameters list.
The custom function listed here is a gaussian root depth distribution.
}
