% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillDepthsUsingSimpleCTM.R
\name{fillDepthsUsingSimpleCTM}
\alias{fillDepthsUsingSimpleCTM}
\title{Fill depth increments using simple Cohort Theory Model.}
\usage{
fillDepthsUsingSimpleCTM(totalDepth = 150, rootDepthMax = 30,
  totalRootBmass = 3000, refractoryFrac = 0.2, bgTurnover = 0.5,
  ssc = 20, depthBelowMHW = 10, omDecayRate = 0.8, coreYear = 2015,
  rootShape = "linear", k1 = 0.085, k2 = 1.99, extraStatsOn = F,
  graphingOn = F, plotName = "rCtmOutput")
}
\arguments{
\item{totalDepth}{}

\item{rootDepthMax}{Depth of the root zone in cm below surface.}

\item{totalRootBmass}{Total Root Biomass, g/m2.}

\item{refractoryFrac}{Refractory Portion of organic production, g/g.}

\item{bgTurnover}{Below Ground Turnover Rate, 1/yr.}

\item{ssc}{Suspended Sediment Concentration, mg per liter.}

\item{depthBelowMHW}{Depth of Marsh Surface Below Mean High Water.}

\item{omDecayRate}{Labile organic matter decay rate, 1/yr.}

\item{coreYear}{Core Collection Year CE.}

\item{rootShape}{Could be linear, exponential, or custom.}

\item{k1}{Organic and Inorganic Self Packing Densities: k1 and k2.}

\item{k2}{}

\item{extraStatsOn}{}

\item{graphingOn}{}

\item{plotName}{}
}
\description{
This function fills 1 cm depth increments down to a maximum depth interval 
using a simple version of Cohort Theory Model. It takes several biotic and 
abiotic inputs, and fills cells assuming steady state conditions.
It outputs a table of depth increments with root mass, refractory 
and labile organic matter, inorganic matter, and age-depth relationships.
There are optionalities to create derivative statistics used to simulate
measurements and to graph the outputs.
}
