% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodTimeFromDatum.R
\name{floodTimeFromDatum}
\alias{floodTimeFromDatum}
\title{Calculate flood time from datum}
\usage{
floodTimeFromDatum(z, datumHigh, datumLow)
}
\arguments{
\item{z}{a numeric, elevation of the marsh}

\item{datumHigh}{a numeric, highest water level of the tidal period}

\item{datumLow}{a numeric, lowest water level of the tidal period}
}
\value{
a numeric, flood time in hours per tidal period
}
\description{
This function calculates flood time in hours from an elevation, and a high and low tidal datum.
}
\references{
Hickey et al In Press. Tidal inundation modeling within GIS.

Department of the Navy. Hydrographic Branch. 1994. Australian National Tide Tables: Australia, Papua & New Guinea, Australian Hydrographic Publication 11.
}
