% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodTimeLinear.R
\name{floodTimeLinear}
\alias{floodTimeLinear}
\title{Calculate flood time using linear method}
\usage{
floodTimeLinear(z, datumHigh, datumLow, tidalCycleLength = 1)
}
\arguments{
\item{z}{a numeric, elevation of the marsh}

\item{datumHigh}{a numeric, highest water level of the tidal period}

\item{datumLow}{a numeric, lowest water level of the tidal period}

\item{tidalCycleLength}{a numeric, the length of time of the tidal cycle. This is set to 1 to esimate flood time as a fraction of the tidal cycle.}
}
\value{
a numeric, flood time per tidal period
}
\description{
This function calculates flood time in hours from an elevation, and a high and low tidal datum.
}
