% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodTimeTrig.R
\name{floodTimeTrig}
\alias{floodTimeTrig}
\title{Calculate flood time using trigonometric method}
\usage{
floodTimeTrig(z, datumHigh, datumLow, tidalCycleLength = 1)
}
\arguments{
\item{z}{a numeric, elevation of the marsh}

\item{datumHigh}{a numeric, highest water level of the tidal period}

\item{datumLow}{a numeric, lowest water level of the tidal period}

\item{tidalCycleLength}{a numeric, the length of time of the tidal cycle. This is set to 1 so that it esimates fractional flood time}
}
\value{
a numeric, flood time per tidal period
}
\description{
This function calculates flood time in hours from an elevation, and a high and low tidal datum.
}
\references{
Hickey, R. (2019). Tidal inundation modeling within GIS. Journal of Coastal Conservation, 23(3), 599-606.

Department of the Navy. Hydrographic Branch. 1994. Australian National Tide Tables: Australia, Papua & New Guinea, Australian Hydrographic Publication 11.
}
