% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisCs137Scenarios.R
\name{listCs137Scenarios}
\alias{listCs137Scenarios}
\title{List available 137Cs fallout scenarios}
\usage{
listCs137Scenarios()
}
\value{
Character vector of scenario names
}
\description{
Shows the built-in fallout scenarios included with the package.
}
