% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.RCMEM.R
\name{model2netcdf.RCMEM}
\alias{model2netcdf.RCMEM}
\title{Function to convert RCMEM model output to standard netCDF format}
\usage{
model2netcdf.RCMEM(
  outdir = "models/rcmem/demo_run/input_demo_out/out/ENS-00001-SERC/",
  sitelat = 38.874544,
  sitelon = -76.548628,
  start_date,
  end_date,
  delete_raw = FALSE
)
}
\arguments{
\item{outdir}{Location of RCMEM model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start year of the simulation}

\item{end_date}{End year of the simulation}

\item{delete_raw}{logical: remove out.csv after converting?}
}
\description{
Convert RCMEM output to netCDF
}
\details{
Converts all output contained in a folder to netCDF.
}
\author{
J. Holmquist
}
