% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEcAn.RCMEM-package.R
\docType{package}
\name{PEcAn.RCMEM-package}
\alias{PEcAn.RCMEM}
\alias{PEcAn.RCMEM-package}
\title{PEcAn.RCMEM-package}
\description{
PEcAn support for the tidal biochemistry model RCMEM
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pecanproject.github.io}
  \item Report bugs at \url{https://github.com/PecanProject/pecan/issues}
}

}
\author{
\strong{Maintainer}: James Holmquist \email{HolmquistJ@si.edu}

Authors:
\itemize{
  \item Abigail Lewis
}

}
\keyword{internal}
