% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictedBiomass.R
\name{predictedBiomass}
\alias{predictedBiomass}
\title{Predict aboveground biomass}
\usage{
predictedBiomass(z = 0, bMax = 2500, zVegMax = 3, zVegMin = -1,
  zVegPeak = NA)
}
\arguments{
\item{z}{marsh elevation}

\item{bMax}{maximum biomass}

\item{zVegMax}{upper elevation of biomass limit}

\item{zVegMin}{lower elevation of biomass limit}

\item{zVegPeak}{(optional) elevation of peak biomass}
}
\value{
a numerical value for the aboveground biomass
}
\description{
Function takes elevation (z) as an input and outputs biomass.
Maximum growing biomass (bmax), upper and lower elevation limits
(zVegMin and zVegMax) are mandatory parameters
The elevation of peak biomass is an optional parameter.
This is a version of the function that can create skewed parabolic distributions.
}
\examples{
predictedBiomass(0, 2500, 3, -1)
}
