% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLunarNodalCycle.R
\name{predictLunarNodalCycle}
\alias{predictLunarNodalCycle}
\title{Predict lunar nodal cycle}
\usage{
predictLunarNodalCycle(year, floodElv, MSL0, MSL, lunarNodalAmp)
}
\arguments{
\item{floodElv}{a numeric, high tide level over the last complete 18 year datum period}

\item{MSL0}{a numeric, mean sea-level over the last complete 18 year datum period}

\item{MSL}{a vector, mean sea-level for each scenario we are predicting}

\item{lunarNodalAmp}{the amplitude of the 18 year lunar nodal cycle}
}
\description{
This function builds a high tide level based on the 18.1 year lunar nodal cycle.
}
\references{
Morris, J. T., Sundberg, K., & Hopkinson, C. S. (2013). Salt marsh primary production and its responses to relative sea level and nutrients in estuaries at Plum Island, Massachusetts, and North Inlet, South Carolina, USA. Oceanography, 26(3), 78-84.

Rochlin, I., & Morris, J. T. (2017). Regulation of salt marsh mosquito populations by the 18.6‐yr lunar‐nodal cycle. Ecology, 98(8), 2059-2068.
}
