% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCohortMem2.R
\name{runCohortMem2}
\alias{runCohortMem2}
\title{Run the cohort MEM model given a sea level rise scenario, initial elevation, a set of plant trait and geopohysical parameters, and optionally, an initial set of soil mass cohorts.}
\usage{
runCohortMem2(
  run_spinup = 1,
  run_scenario = 1,
  spinup_iterations = 10,
  years_per_spinup_iter = 10,
  competition_function = 0,
  scenario_calendar_years = 1928:2018,
  scenario_iterations = length(scenario_calendar_years),
  years_per_scenario_iter = 1,
  initElv = -11,
  initial_cohort_vol = 100,
  initial_cohort_years = 99999,
  initial_scenario = NA,
  initial_cohorts = NA,
  initial_species = NA,
  bMax = c(0.0522, 0.0522),
  zVegMin = c(-1.4, -1.4 + 0.1),
  zVegMax = c(1.8, 1.8 + 0.1),
  zVegPeak = c(1.5, 1.5 + 0.1),
  rootToShoot = c(1.636903, 1.636903),
  rootTurnover = c(0.5, 0.5),
  abovegroundTurnover = c(1, 1),
  rootDepthMax = c(48, 48),
  rootPackingDensity = c(0.07, 0.07),
  species_codes = c("SCAM", "SPPA"),
  meanOmPackingDensity = 0.1,
  meanMineralPackingDensity = 1.5,
  omDecayRateFast = 0.99,
  omDecayRateSlow = 0.001,
  recalcitrantFrac = 0.4,
  captureRate = 5,
  suspendedSediment = 6 * 1e-06,
  rootShape = c(0, 0),
  relTol = 1e-06,
  safe_mode = T,
  flood_frequency = c(0.5, 0.46497542, 0.03502458) * 705.79,
  msl,
  mhwMat,
  mlwMat
)
}
\arguments{
\item{run_spinup}{binary 1 or 0, 1 runs a spinup, 0 skips over a spinup an initiates using a separate initial set of soil cohorts}

\item{run_scenario}{bionary 1 or 0, 1 runs a sea-level rise scnario, 0 only runs a spinup}

\item{spinup_iterations}{integer, the number of spinup iterations to run}

\item{years_per_spinup_iter}{integer, the number of years of marsh formation to aggregate into each spinup iteration}

\item{competition_function}{binary 1 or 0, 0 indicates a winner take all plant competition scenario, 1 indicates a coexistence with competition scenario}

\item{scenario_calendar_years}{a numeric vector of calendar years}

\item{scenario_iterations}{integer, the number of iterations in the sea-level change scenario}

\item{years_per_scenario_iter}{integer, the number of years included in each iteration of the sea-level rise scenario}

\item{initElv}{numeric, initial elevation of the scenario stage}

\item{initial_cohort_vol}{for the intial cohort, cohort 1, the initial volume, generically set to 1 m deep}

\item{initial_cohort_years}{for the intial cohort, cohort 1, the number of years of formation. This is arbitrarily set to a high number}

\item{initial_scenario}{data frame, a scenario table from a previous rCMEM run}

\item{initial_cohorts}{data frame, a cohort table from a previous rCMEM run}

\item{initial_species}{data frame, a species table from a previous rCMEM run}

\item{bMax}{a numeric vector, for each species a peak biomass in grams per square centimeters}

\item{zVegMin}{a numeric vector, for each species minimum growing elevation in relative tidal elevation (elevation - msl) / (mhw - msl)}

\item{zVegMax}{a numeric vector, for each species maximum growing elevation in relative tidal elevation (elevation - msl) / (mhw - msl)}

\item{zVegPeak}{a numeric vector, for each species peak growing elevation in relative tidal elevation (elevation - msl) / (mhw - msl)}

\item{rootToShoot}{a numeric vector, for each species, the belowground to aboveground biomass ratio}

\item{rootTurnover}{a numeric vector, for each species, the belowground annual turnover rate}

\item{abovegroundTurnover}{a numeric vector, for each species, the aboveground annual turnover rate}

\item{rootDepthMax}{a numeric vector, for each species, the maximum rooting depth in centimeters below the surface}

\item{rootPackingDensity}{a numeric vector, for each species, the root self packing density in grams per square centimeters}

\item{species_codes}{a character vector, for each species, a unique code or name to track species in output tables}

\item{meanOmPackingDensity}{numeric, mean soil organic matter self packing density in grams per square centimeters}

\item{meanMineralPackingDensity}{numeric, mean soil mineral matter self packing density in grams per square centimeters}

\item{omDecayRateFast}{numeric, annual mass lost of the fast decaying portion of soil organic matter}

\item{omDecayRateSlow}{numeric, annual mass lost of the slow decaying portion of soil organic matter}

\item{recalcitrantFrac}{numeric, fraction of soil organic matter in the slow decaying pool}

\item{captureRate}{numeric, the number of times that a column of sediment will settle out in a full tidal cycle}

\item{suspendedSediment}{numeric, suspended sediment concentration in grams per cubic centimeter}

\item{rootShape}{binary indicating the shape of the rooting zone, 0 for linear, 1 for exponential shape}

\item{relTol}{numeric, relative tolerance of the non-linear equation solving algorithm}

\item{safe_mode}{true or false, turns on safe_mode which will stop code and throw errors if some incongrous inputs are made}

\item{flood_frequency}{numeric vector, indicates the frequence of flood events, must be the same length as the widths of mhwMat and mlwMat}

\item{msl}{numeric vector, the mean sea levels of each stage of a sea-level change scenario}

\item{mhwMat}{numeric matrix, the levels of annual average flood events for each stage of the sea level change scenario}

\item{mlwMat}{numeric matrix, the levels of annual average ebb events for each stage of the sea level change scenario}
}
\value{
a list with three data frames, scenario: an annual time series tracking annual inputs and outputs, cohorts: a long form table tracking mass cohorts and soil volumes, and species: species level predictions of
}
\description{
Run the cohort MEM model given a sea level rise scenario, initial elevation, a set of plant trait and geopohysical parameters, and optionally, an initial set of soil mass cohorts.
}
