% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMemWithCohorts.R
\name{runMemWithCohorts}
\alias{runMemWithCohorts}
\title{Run Marsh Equilibrium Model with Cohorts}
\usage{
runMemWithCohorts(startYear, endYear = startYear + 99, rslrT1, rslrTotal,
  initElv, MSL, MSL0 = MSL[1], MHW, MHHW = NA, MHHWS = NA, ssc,
  lunarNodalAmp, bMax, zVegMin, zVegMax, zVegPeak, plantElevationType,
  rootToShoot, rootTurnover, rootDepthMax, shape = "linear", omDecayRate,
  recalcitrantFrac, settlingVelocity, omPackingDensity = 0.085,
  mineralPackingDensity = 1.99, rootPackingDensity = omPackingDensity,
  coreYear = NA, coreDepth = 100, coreMaxs = 1:coreDepth,
  coreMins = coreMaxs - 1, ...)
}
\arguments{
\item{startYear}{an integer, year in form YYYY, the start year of the scenario}

\item{endYear}{an integer, year in form YYYY, the end year of the scenario}

\item{rslrT1}{a numeric, initial rate of relative sea-level rise}

\item{rslrTotal}{a numeric, total relative sea-level rise over the course of the scanario}

\item{initElv}{a numeric, the initial elevation of the marsh at the start of the scenario}

\item{MSL}{a numeric or a vector, Mean Sea Level at the start of the scenario, or a vector of Mean Sea Levels the same length as the number of years in a scenario}

\item{MSL0}{a numeric, Mean Sea level over the last datum period}

\item{MHW}{a numeric, Mean High Water level over the last datum period}

\item{MHHW}{a numeric (optional), Mean Higher High Water level over the last datum period}

\item{MHHWS}{a numeric (optional), Mean Higher High Spring Tide Water level over the last datum period}

\item{ssc}{a numeric, suspended sediment concentration of the water column}

\item{lunarNodalAmp}{a numeric, the amplitude of the 18-year lunar nodal cycle}

\item{bMax}{a numeric, maximum biomass}

\item{zVegMin}{a numeric, lower elevation of biomass limit}

\item{zVegMax}{a numeric, upper elevation of biomass limit}

\item{zVegPeak}{a numeric, (optional) elevation of peak biomass}

\item{plantElevationType}{character, either "orthometric" or "dimensionless", specifying elevation reference of the vegetation growing elevations}

\item{rootToShoot}{a numeric, root to shoot ratio}

\item{rootTurnover}{a numeric, below ground biomass annual turnover rate}

\item{rootDepthMax}{a numeric, maximum (95\%) rooting depth}

\item{shape}{a character, "linear" or "exponential" describing the shape of the relationship between depth and root mass}

\item{omDecayRate}{a numeric, annual fractional mass lost}

\item{recalcitrantFrac}{a numeric, fraction of organic matter resistant to decay}

\item{settlingVelocity}{a numeric, the number of times a water column will clear per tidal cycle}

\item{omPackingDensity}{a numeric, the bulk density of pure organic matter}

\item{mineralPackingDensity}{a numeric, the bulk density of pure mineral matter}

\item{rootPackingDensity}{a numeric, the bulk density of pure root matter}

\item{coreYear}{an integer, year in form YYYY, (optional) specify a year to simulate taking a sediment core}

\item{coreDepth}{an integer, depth, (optional) specify a depth to simulate coring to and assume 1 cm sampling intervals}

\item{coreMaxs}{a vector of sampling depth maximums (optional) to simulate coring subsables, this is an alternative to depth, and 1cm increments}

\item{coreMins}{a vector of sampling depth minimums (optional) to simulate coring subsables, this is an alternative to depth, and 1cm increments}

\item{...}{}
}
\value{
a list of data frames, including the annualized summaries, mapped cohorts tracked for every years, and if core year is specified, a core.
}
\description{
This function takes an initial elevation, average annual suspended sediment concentrtion and tidal properties, plant traits, and decay rates, then builds a scenario, and runs the marsh equilibrium model over that scenario tracking individual sediment mass cohorts including minderal, live root, slow decay organic mater pool and a fast organic matter pool.
}
