% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runToEquilibrium.R
\name{runToEquilibrium}
\alias{runToEquilibrium}
\title{Run to equilibrium with proscribed surface mineral input}
\usage{
runToEquilibrium(
  minAge = 50,
  maxAge = 12000,
  minDepth,
  recordEvolution = FALSE,
  relTol = 1e-06,
  absTol = 1e-08,
  ...
)
}
\arguments{
\item{minAge}{run the model for at least these many years}

\item{maxAge}{do not run the model for longer then these many years}

\item{minDepth}{run the model to create a profile at least this deep}

\item{recordEvolution}{a boolean flag to record how the equilibrium profile evolves}

\item{relTol}{stop the evolution if the relative difference between the two oldest cohorts is less then this}

\item{absTol}{stop the evolution if the absolute difference between the two oldest cohorts is less then this}

\item{...}{arguments to be passed to \code{addCohort}}
}
\value{
a data frame by age cohort with the layer top and bottom depths, age, and mass of the fast organic, slow organic, mineral, and root biomass pools OR a list that is a record of age cohorts
}
\description{
Repeatedly add new cohorts to a soil profile until the oldest layers are not significantly different AND the oldest layer are above some minimum age, within some upper bound of soil age.
}
