% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedimentInputs.R
\name{sedimentInputs}
\alias{sedimentInputs}
\title{Annual mineral inputs into marsh surface}
\usage{
sedimentInputs(ssc, meanTidalHeight, nTidesPerYear = 704,
  soilLength = 1, soilWidth = 1, ...)
}
\arguments{
\item{ssc}{a numeric representing suspended sediment concentration in (mg per liter)}

\item{meanTidalHeight}{annual mean tide above marsh elevation}

\item{nTidesPerYear}{number of tides per year}

\item{soilLength}{unit length of interest}

\item{soilWidth}{unit width of interest}

\item{...}{}
}
\value{
a numeric that is the mass of mineral added in one year to the top of the marsh
}
\description{
This function cacluates the annaual sediment input into the marsh surface via the following formula:
Number of tides per year, times the unit area of interest, times the mean tidal volume above the marsh surface in cm$^3$, times the suspended sediment concentration in mg per liter, times 1e-6 conversion factor to convert from liter to cm$^3, resulting in the mineral mass added per year. The number of times (nTidesPerYear), unit area length (soilLength), and unit area width (soilWidth) are included in the consts list (see specific names).
}
