% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFluxes.R
\name{simulateFluxes}
\alias{simulateFluxes}
\title{Simulate Greenhouse Gas Fluxes}
\usage{
simulateFluxes(
  cohorts,
  scenario,
  omToOcParams = list(B0 = 0, B1 = 0.48),
  salinity = 32,
  salThreshold = 14,
  salRate = 0.5,
  maxCH4effect = 1,
  agbTurnoverRate = NA
)
}
\arguments{
\item{agbTurnoverRate}{}
}
\value{

}
\description{
This function takes outputs from runMemWithCohorts and uses them to simulate several annualized greenhouse gas flux variables
}
