% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRadioisotopeDates.R
\name{simulateRadioisotopeDates}
\alias{simulateRadioisotopeDates}
\title{Simulate radioisotop dates}
\usage{
simulateRadioisotopeDates(
  core,
  pb210_influx_Bq_m2_yr = 200,
  pb210_supported_Bq_kg = 19.8,
  cs137_scenario = "cs137_N_non_europe",
  cs137_anomaly = 0.6224604,
  cs137_baseline_Bq_m2_yr = NA,
  cs137_erosion_loss = 0.5,
  pb210_erosion_loss = 0.01,
  cs137_mobility_factor = 3,
  pb210_mobility_factor = 2
)
}
\arguments{
\item{core}{dataframe, output from build soil core}

\item{pb210_influx_Bq_m2_yr}{numeric, atmospheric flux in Bequerels per kilogram per square meter}

\item{pb210_supported_Bq_kg}{numeric, Concnetration of supported 210Pb in Bequerels per gram}

\item{cs137_scenario}{data frame with a year column indicating calendar years, and a second column indicating cs137 influx in Bequerels per square cm}

\item{cs137_anomaly}{numeric, fractional anomoly inidicating site offset from regional scenario}

\item{cs137_baseline_Bq_m2_yr}{numeric, baseline fallout of cs137 without aboveground nuclear wepons testing. Defaults to the minimum value in the 137Cs fallout scenario}

\item{cs137_erosion_loss}{numeric, bound between 1 and 0, how much of the annual influx of 137Cs is lost to erosion}

\item{pb210_erosion_loss}{numeric, bound between 1 and 0, how much of the annual influx of 210Pb is lost to erosion}

\item{cs137_mobility_factor}{integer or zero, indicates how many adjacent depth increments will exchange 137Cs when simulating mobility}

\item{pb210_mobility_factorinteger}{or zero, indicates how many adjacent depth increments will exchange 210Pb when simulating mobility}
}
\value{
a data frame with the same structure as core input, but with additional
}
\description{
Simulate radioisotop dates
}
