% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSetData.R
\name{sumulateSetData}
\alias{sumulateSetData}
\title{Simulate SET Data}
\usage{
sumulateSetData(cohorts, scenario, markerHorizonYear = NA)
}
\arguments{
\item{cohorts}{}

\item{markerHorizonYear}{}
}
\value{
a dataframe of net elevation changes and accretion rates relative to year
}
\description{
This will take MEM outputs and simulate sediment elevation table and marker horizon data.
}
