% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimCohorts.R
\name{trimCohorts}
\alias{trimCohorts}
\title{Trim cohorts}
\usage{
trimCohorts(cohorts)
}
\arguments{
\item{cohorts}{data frame tracking soil cohort mass pools}
}
\description{
Function cleans up no data values from cohorts table
}
