% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.RCMEM.R
\name{write.config.RCMEM}
\alias{write.config.RCMEM}
\title{Write RCMEM configuration files}
\usage{
write.config.RCMEM(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
configuration file for RCMEM for given run
}
\description{
Writes a RCMEM config file.
}
\details{
Requires a pft xml object, a list of trait values for a single RCMEM run,
and the name of the file to create
}
\author{
Rob Kooper, edited by James Holmquist
}
