% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zStarToZ.R
\name{zStarToZ}
\alias{zStarToZ}
\title{Convert from Z* to elevation}
\usage{
zStarToZ(zStar, MHW, MSL)
}
\arguments{
\item{zStar}{a numeric describing the normalized Z* elevation}

\item{MHW}{a numeric, mean high water tide}

\item{MSL}{a numeric, mean sea level}
}
\value{
a numerical describing the marsh elevation
}
\description{
Convert from Z* to elevation
}
