% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertZToZstar.R
\name{zToZstar}
\alias{zToZstar}
\title{Convert elevation to Z*}
\usage{
zToZstar(z, meanHighWater, meanSeaLevel)
}
\arguments{
\item{z}{a numeric elevation}

\item{meanHighWater}{a numeric, mean high water tide}

\item{meanSeaLevel}{a numeric, mean sea level}
}
\value{
elevation normalized to the tidal range
}
\description{
(z-meanSeaLevel)/(meanHighWater-meanSeaLevel)
}
