% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCohort.R
\name{addCohort}
\alias{addCohort}
\title{Add another cohort to the soil profile}
\usage{
addCohort(massPools, rootTurnover, rootOmFrac, omDecayRate, packing,
  mineralInput_g_per_yr.fn = sedimentInputs,
  mineralInput_g_per_yr = NA, massLiveRoots.fn = massLiveRoots,
  depthOfNonRootVolume.fn = depthOfNonRootVolume, dt_yr = 1, ...)
}
\arguments{
\item{massPools}{A data frame containing the following colums: age, fast_OM, slow_OM, mineral, layer_top, layer_bottom, root_mass. Each row of this data frame represents a single cohort. To increase runtime in simulations, the profile is buffered with NA's at the top of the data frame, see code comments for details.}

\item{rootTurnover}{A numeric as the root turnover time in faction per year.}

\item{rootOmFrac}{A list of numerics called fast and slow which is the allocation fraction between fast and slow organic matter pool respectively for dead roots.}

\item{omDecayRate}{A list of numerics called fast and slow which is the decay rate for the fast and slow organic matter pool in fraction per year.}

\item{packing}{A list of numerics with mineral and organic arguments representing their respective packing densities.}

\item{mineralInput_g_per_yr.fn}{A function that returns the mineral input in grams per year (numeric)}

\item{massLiveRoots.fn}{A function that returns the mass of live roots for specified depth layers, must accept \code{layerBottom} and \code{layerTop} as arguments.}

\item{depthOfNonRootVolume.fn}{A function that returns the depth of a specified volumen of soil, must accept \code{nonRootVolume} as an argument.}

\item{dt_yr}{A numeric of the time step in years. Typically set to 1 year.}

\item{...}{arguments to be passed to the specified functions}
}
\value{
A new data frame similar to massPool with the added cohort and simulated change in the cohort pools.
}
\description{
This function adds a new cohort to a soil profile. Cohorts are constructed as follows:
\enumerate{
       \item Cohorts are aged by dt_yr incrament.
       \item Organic matter pools decay
       \item If there are non-zero mineral inputs, they are added to the top of the soil profile as a new cohort.
       \item The top and bottom of the cohort layer is recalculated.
       \item A new root profile is constructed based on the cohort layers.
}
}
