% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliverSedimentFlexibly.R
\name{deliverSedimentFlexibly}
\alias{deliverSedimentFlexibly}
\title{Deliver sediment flexibly}
\usage{
deliverSedimentFlexibly(z, ssc, MSL, MHW, MHHW = NA, MHHWS = NA,
  settlingVelocity, ...)
}
\arguments{
\item{z}{a numeric, elevation of the marsh}

\item{ssc}{a numeric, suspended sediment concentration of the water column}

\item{MSL}{a numeric, Mean Sea Level}

\item{MHW}{a numeric, Mean High Water level}

\item{MHHW}{a numeric, Mean Higher High Water level}

\item{MHHWS}{a numeric, Mean Higher High Spring Tide Water level}

\item{settlingVelocity}{a numeric, the number of times a water column will clear per tidal cycle}

\item{...}{}
}
\value{
a numeric, sediment delivered in a year of flooding
}
\description{
Function decides which sediment module to use based on available inputs
}
