% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthOfNonRootVolume.R
\name{depthOfNotRootVolume}
\alias{depthOfNotRootVolume}
\title{Depth of the given soil volume}
\usage{
depthOfNotRootVolume(nonRootVolume, totalRootMass_per_area, rootDepthMax,
  parms = NULL, consts)
}
\arguments{
\item{nonRootVolume}{a numerical value that is the volume of interest, generally in cm3.}

\item{totalRootMass_per_area}{an integer that is the total mass per area of the roots, generally in g cm-3}

\item{rootDepthMax}{an integer that is the maximum root depth, generally in cm}

\item{parms}{placeholder currently NULL}

\item{consts}{a list of constants that must include numerics representing the soilLength (cm) and soilWidth (cm) for the area of interest, a flag shape, and a list called packing that includes a value of the root packing (or density).}
}
\description{
Given a shape and mass for the roots, calculate the depth of a specific non-root volume.
}
