% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massLiveRoots.R
\name{massLiveRoots}
\alias{massLiveRoots}
\title{Mass of living roots}
\usage{
massLiveRoots(layerBottom, layerTop, totalRootMass_per_area, rootDepthMax,
  soilLength = 1, soilWidth = 1, shape,
  expDecayRate_perMaxDepth = log(0.05), ...)
}
\arguments{
\item{layerBottom}{an array of depths from soil top to the bottom of the layer, generally in cm}

\item{layerTop}{an array of depths from soil top to the top of the layer, generally in cm}

\item{totalRootMass_per_area}{an integer that is the total mass per area of the roots, generally in g cm-3}

\item{rootDepthMax}{an integer that is the maximum root depth, generally in cm}

\item{soilLength}{unit length of soil area, generally 1}

\item{soilWidth}{unit width of soil area, generally 1.}

\item{shape}{flag of root shape, only \code{linear} is implimented.}

\item{...}{}
}
\value{
an array of the associated live root mass for each layer specified.
}
\description{
This function returns the mass of the living roots between two layers. Given a mass per area of roots and a maximum rooting depth, this calculates the total mass between two layers with a specified distribution. Currently only a linear distribution is implimented. \cr
The linear algorithm is as follows:  \itemize{
   \item #mass_per_depth = slope * depth + intercept
   \item slope <- -2 * totalRootMass / (rootDepthMax^2)
   \item intercept <- 2 * totalRootMass / rootDepthMax
   \item #mass = integral(mass_per_depth, depth)
   \item rootMass <- intercept * (layerBottom-layerTop) + slope/2 * (layerBottom ^2-layerTop^2)
   }
}
\examples{
massLiveRoots(layerBottom = 1:10, layerTop = 0:9, 
              totalRootMass_per_area = 0.3, rootDepthMax=5,
              soilLength=1, soilWidth=1, shape='linear')
}
