% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runToEquilibrium.R
\name{runToEquilibrium}
\alias{runToEquilibrium}
\title{Run to equilbrium with proscribed depth to mean high tide}
\usage{
runToEquilibrium(parms, consts, minAge = 50, maxAge = 200,
  recordEvolution = FALSE, relTol = 1e-06, absTol = 1e-08)
}
\arguments{
\item{parms}{a list of parameters that are passed to addCohort individually}

\item{consts}{a list of constants that are passed to addCohort as a list}

\item{minAge}{run the model for atleast these many years}

\item{maxAge}{do not run the model for longer then these many years}

\item{recordEvolution}{a boolean flag to record how the equalibrium profile evolves}

\item{relTol}{stop the evolution if the relative difference between the two oldest cohorts is less then this}

\item{absTol}{stop the evolution if the absolute difference between the two oldest cohorts is less then this}
}
\value{
a data frame by age cohort with the layer top and bottom depths, age, and mass of the fast organic, slow organic, mineral, and root biomass pools.
}
\description{
Run to equilbrium with proscribed depth to mean high tide
}
