% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFluxes.R
\name{simulateFluxes}
\alias{simulateFluxes}
\title{Simulate Greenhouse Gas Fluxes}
\usage{
simulateFluxes(cohorts, scenario, omToOcParams = list(B0 = 0, B1 = 0.48),
  salinity = 32, salThreshold = 18, salRate = 0.5,
  maxCh4ToCo2 = 1/5)
}
\arguments{
\item{maxCH4effect}{}
}
\value{
list of 2 data frames, one the input cohorts table with depthwise fluxes added, the second the annualized scenario table with annualized fluxes added
}
\description{
This function takes outputs from runMemWithCohorts and uses them to simulate several annualized greenhouse gas flux variables
}
