% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSetData.R
\name{simulateSetData}
\alias{simulateSetData}
\title{Simulate SET Data}
\usage{
simulateSetData(cohorts, scenario, markerHorizonYear = NA)
}
\arguments{
\item{cohorts}{a data frame, ouptut from runCohortMem, tracking mineral and organic mass cohorts over each year of the simulation}

\item{scenario}{a data frame, second ouptut from runCohortMem, tracking annualized summaries of drivers and marsh profiles dynamics}

\item{markerHorizonYear}{an integer, the year a marker horizon was lain down, that were using to simulate an accretion measurement}
}
\value{
a dataframe of net elevation changes and accretion rates relative to year
}
\description{
This will take MEM outputs and simulate sediment elevation table and marker horizon data.
}
