% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zToZstar.R
\name{zToZstar}
\alias{zToZstar}
\title{Convert elevation to Z*}
\usage{
zToZstar(z, MHW, MSL)
}
\arguments{
\item{z}{a numeric elevation}

\item{MHW}{a numeric, mean high water tide}

\item{MSL}{a numeric, mean sea level}
}
\value{
elevation normalized to the tidal range
}
\description{
(z-MSL)/(MHW-MSL)
}
